API update and HTTP request overhaul
The API link has been updated to work with PIcarto's new API format.
The way that the bot checks to see if a streamer has come online has
been overhauled again. This time the focus was on reducing the number of
HTTP requests and JSON parse calls to improve scaling.

Minor code cleanup and adjustments made elsewhere, mostly removing
unused code and unused if/else blocks.