Updating with a note about the 5/6/18 issues
The bot is not able to maintain a stable CPU usage, and it is crushing the server it is hosted on. There was a sudden spike on May 6th, at about 4PM UTC. I do not yet know what is causing it, though the API the bot reads from Picarto has also been having serious issues. 502 errors were the normal response for about 12 hours. I deployed a hotfix to prevent the bot from being bogged down by the errors, but either the fix did not work or there is another more in-depth issue. I suspect the information returned by the API request was increased, because the time the bot spent Parsing the data into a JSON went up by a factor of 5, but I have not yet been able to test it. Once the 502 errors subside (and I have some time, sorry) I will investigate further.